Created library for chapter examples
