+++ .gitignore
@@ -1,3 +1,3 @@
 /target
-examples/image.png
+chapter_code/image.png
 .idea
