+++ Cargo.toml
@@ -10,10 +10,3 @@ lazy_static = "1.1"
 mustache = "0.9"
 pulldown-cmark = "0.9.1"
 rouille = "3.0.0"
-
-[dev-dependencies]
-vulkano = "0.28.0"
-vulkano-shaders = "0.28.0"
-image = "0.24.0"
-winit = "0.26.0"
-vulkano-win = "0.28.0"
