+++ chapter_code/src/bin/compute_pipeline.rs
