+++ chapter_code/src/bin/images/image_clear.rs
@@ -7,27 +7,22 @@
 // notice may not be copied, modified, or distributed except
 // according to those terms.
 
-//! This is the source code of the "Drawing a fractal with a compute shader" subchapter
-//! from the "Using images" chapter at http://vulkano.rs.
+//! This is the source code of the first three subchapters from the "Using images" chapter at http://vulkano.rs.
 //!
 //! It is not commented, as the explanations can be found in the guide itself.
 
-use image::ImageBuffer;
-use image::Rgba;
+use image::{ImageBuffer, Rgba};
 use vulkano::buffer::{BufferUsage, CpuAccessibleBuffer};
 use vulkano::command_buffer::{AutoCommandBufferBuilder, CommandBufferUsage};
-use vulkano::descriptor_set::{PersistentDescriptorSet, WriteDescriptorSet};
 use vulkano::device::{physical::PhysicalDevice, Device, DeviceExtensions, Features};
-use vulkano::format::Format;
-use vulkano::image::{view::ImageView, ImageDimensions, StorageImage};
+use vulkano::format::{ClearValue, Format};
+use vulkano::image::{ImageDimensions, StorageImage};
 use vulkano::instance::{Instance, InstanceExtensions};
-use vulkano::pipeline::Pipeline;
-use vulkano::pipeline::{ComputePipeline, PipelineBindPoint};
 use vulkano::sync;
 use vulkano::sync::GpuFuture;
 use vulkano::Version;
 
-fn main() {
+pub fn main() {
     let instance = Instance::new(None, Version::V1_1, &InstanceExtensions::none(), None)
         .expect("failed to create instance");
 
@@ -52,72 +47,16 @@ fn main() {
 
     let queue = queues.next().unwrap();
 
-    mod cs {
-        vulkano_shaders::shader! {
-            ty: "compute",
-            src: "
-#version 450
-
-layout(local_size_x = 8, local_size_y = 8, local_size_z = 1) in;
-
-layout(set = 0, binding = 0, rgba8) uniform writeonly image2D img;
-
-void main() {
-    vec2 norm_coordinates = (gl_GlobalInvocationID.xy + vec2(0.5)) / vec2(imageSize(img));
-
-    vec2 c = (norm_coordinates - vec2(0.5)) * 2.0 - vec2(1.0, 0.0);
-
-    vec2 z = vec2(0.0, 0.0);
-    float i;
-    for (i = 0.0; i < 1.0; i += 0.005) {
-        z = vec2(
-            z.x * z.x - z.y * z.y + c.x,
-            z.y * z.x + z.x * z.y + c.y
-        );
-
-        if (length(z) > 4.0) {
-            break;
-        }
-    }
-
-    vec4 to_write = vec4(vec3(i), 1.0);
-    imageStore(img, ivec2(gl_GlobalInvocationID.xy), to_write);
-}"
-        }
-    }
-
-    let shader = cs::load(device.clone()).expect("failed to create shader module");
-
-    let compute_pipeline = ComputePipeline::new(
-        device.clone(),
-        shader.entry_point("main").unwrap(),
-        &(),
-        None,
-        |_| {},
-    )
-    .expect("failed to create compute pipeline");
-
+    // Image creation
     let image = StorageImage::new(
         device.clone(),
         ImageDimensions::Dim2d {
             width: 1024,
             height: 1024,
-            array_layers: 1,
+            array_layers: 1, // images can be arrays of layers
         },
         Format::R8G8B8A8_UNORM,
         Some(queue.family()),
-    )
-    .unwrap();
-
-    let view = ImageView::new(image.clone()).unwrap();
-    let layout = compute_pipeline
-        .layout()
-        .descriptor_set_layouts()
-        .get(0)
-        .unwrap();
-    let set = PersistentDescriptorSet::new(
-        layout.clone(),
-        [WriteDescriptorSet::image_view(0, view.clone())], // 0 is the binding
     )
     .unwrap();
 
@@ -136,18 +75,10 @@ void main() {
     )
     .unwrap();
     builder
-        .bind_pipeline_compute(compute_pipeline.clone())
-        .bind_descriptor_sets(
-            PipelineBindPoint::Compute,
-            compute_pipeline.layout().clone(),
-            0,
-            set,
-        )
-        .dispatch([1024 / 8, 1024 / 8, 1])
+        .clear_color_image(image.clone(), ClearValue::Float([0.0, 0.0, 1.0, 1.0]))
         .unwrap()
         .copy_image_to_buffer(image.clone(), buf.clone())
         .unwrap();
-
     let command_buffer = builder.build().unwrap();
 
     let future = sync::now(device.clone())
@@ -158,6 +89,7 @@ void main() {
 
     future.wait(None).unwrap();
 
+    // Exporting the result
     let buffer_content = buf.read().unwrap();
     let image = ImageBuffer::<Rgba<u8>, _>::from_raw(1024, 1024, &buffer_content[..]).unwrap();
     image.save("image.png").unwrap();
